Adds rudimentary support for the Gamepad API.
This addresses https://github.com/bfirsh/jsnes-web/issues/19 and is
based off of a prototype I made over four years ago:
https://github.com/bolinfest/jsnes/commit/91daafef70e763fb4ffd6efdcc33c52504a8c421.

On startup, the webapp checks to see if there are any connected
Gamepads. If so, it elects to use a `GamepadController` instead of a
`KeyboardController`. If the Gamepads get disconnected over the
lifetime of the webapp, the current implementation waits for the
Gamepads to be reconnected, though it would be more elegant if it fell
back on the `KeyboardController` in this case.

Note the mapping of the buttons is currently hardcoded. Ideally, there
would be a GUI with a picture of the controller, highlighting one
button at a time, giving the user the chance to press the corresponding
button on their controller to establish the mapping. As a stopgap, we
could also let users do this programmatically by stuffing values in
`localStorage` because it will probably be a bit of work to create the
GUI.