Refactor RunPage to detangle emulator
This allows embedding just the emulator as a React component by
passing ROM data as a string to Emulator. It handles writing to
canvas, sound, keyboard events, mouse events, etc etc.

Minimal viable emulator:

```javascript
<Emulator romData="..." />
```

Boom.

There is still some refactoring to do to clean up keyboard/gamepad
customisation. See also #149.

Also, to make this work we need to release jsnes-web as a package
on npm and document how to embed it.