Rewrite timing module
The timing module is rewritten to run at NTSC 60.098fps.

When requestAnimationFrame is triggered, the given time is rounded
down to the time of the latest frame that needs to be rendered. This
time is then used to determine how many frames the emulator is behind.

When the frame timer falls behind, frames are interlaced between the current
time and the time of the next frame render. For example, if there
is 6ms until the next frame render, and two extra frames are required,
they are generated at +2ms and +4ms.

The benefit of this approach is that it accommodates ANY framerate
given by requestAnimationFrame

See #15